8273838: Enhanced BMP processing
Backport-of: f5cc6f23a46bee06731fdd2b257c8ea8007eebad

+++ src/java.desktop/share/classes/com/sun/imageio/plugins/bmp/BMPImageReader.java
@@ -615,7 +615,7 @@ public class BMPImageReader extends ImageReader implements BMPConstants {
         }
 
         if (metadata.compression == BI_RGB) {
-            long imageDataSize = (width * height * (bitsPerPixel / 8));
+            long imageDataSize = ((long)width * height * (bitsPerPixel / 8));
             if (imageDataSize > (bitmapFileSize - bitmapOffset)) {
                 throw new IIOException(I18N.getString("BMPImageReader9"));
             }
